pyproject.toml
[project]
name = "icon-kitchen-gitlab-sync"
version = "0.1.0"
description = "Generate icon.kitchen URLs, download icons and sync them as GitLab project avatars."
authors = [
    {name = "Daniil Pavlovich",email = "layred.dota2@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright (>=1.58.0,<2.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "pillow (>=12.1.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
icon-kitchen-gitlab-sync = "icon_kitchen_gitlab_sync.cli:main"
